accuracy calculated properly

+++ src/SimpleRhythmGame/Scores.java
@@ -16,7 +16,7 @@ public class Scores {
 		this.maxCombo = maxCombo;
 		this.notesHit = notesHit;
 		this.notesMissed = notesMissed;
-		this.accuracy = notesHit == 0 ? 0 : (notesHit / (notesHit + notesMissed)) * 100;
+		this.accuracy = notesHit == 0 ? 0 : ((float) notesHit / (float) (notesHit + notesMissed)) * 100f;
 	}
 
 	public int getScore() {
